Add redirect to end of login process.
Without this redirect, if the "URL to redirect to after login" was set
to "/user" the returning user would see the login form after logging
in via Shib.  If they, then hit "Reload", the wold see their user page
(as expected).  This redirect allows the login to work correctly.
It may be better to do the login via a controller, but this does now
work as desired.